// jweixin-1.2.0.d.ts
declare namespace Wx {
  /****************************************************************************
   * Type definations
   ****************************************************************************/
  type JSApis =
    | 'onMenuShareTimeline'
    | 'onMenuShareAppMessage'
    | 'onMenuShareQQ'
    | 'onMenuShareWeibo'
    | 'onMenuShareQZone'
    | 'startRecord'
    | 'stopRecord'
    | 'onVoiceRecordEnd'
    | 'playVoice'
    | 'pauseVoice'
    | 'stopVoice'
    | 'onVoicePlayEnd'
    | 'uploadVoice'
    | 'downloadVoice'
    | 'chooseImage'
    | 'previewImage'
    | 'uploadImage'
    | 'downloadImage'
    | 'translateVoice'
    | 'getNetworkType'
    | 'openLocation'
    | 'getLocation'
    | 'hideOptionMenu'
    | 'showOptionMenu'
    | 'hideMenuItems'
    | 'showMenuItems'
    | 'hideAllNonBaseMenuItem'
    | 'showAllNonBaseMenuItem'
    | 'closeWindow'
    | 'scanQRCode'
    | 'chooseWXPay'
    | 'openProductSpecificView'
    | 'addCard'
    | 'chooseCard'
    | 'openCard'
    | 'getLocalImgData';
// ...
  type scanTypes =
    | 'qrCode' // 二维码
    | 'barCode'; // 一维码
  type ImageSize = 'original' | 'compressed';
  type SourceType = 'album' | 'camera';

  /****************************************************************************
   * Interface definations
   ****************************************************************************/

  interface ConfigOptions {
    /**
     * 开启调试模式,调用的所有api的返回值会在客户端alert出来，若要查看传入的参数，
     * 可以在pc端打开，参数信息会通过log打出，仅在pc端时才会打印。
     */
    debug?: boolean;

    /**
     * 公众号的唯一标识
     */
    appId: string;

    /**
     * 生成签名的时间戳
     */
    timestamp: string;

    /**
     * 生成签名的随机串
     */
    nonceStr: string;

    /**
     * 签名
     */
    signature: string;

    /**
     * 需要使用的JS接口列表
     */
    jsApiList: JSApis[];
  }

  /**
   * 所有接口通过wx对象(也可使用jWeixin对象)来调用，参数是一个对象，
   * 除了每个接口本身需要传的参数之外，还有以下通用参数
   */
  interface Callbacks {
    /**
     * 接口调用成功时执行的回调函数。
     */
    success?: (res: any) => void;

    /**
     * 接口调用失败时执行的回调函数。
     */
    fail?: (res: any) => void;

    /**
     * 接口调用完成时执行的回调函数，无论成功或失败都会执行。
     */
    complete?: (res: any) => void;

    /**
     * 用户点击取消时的回调函数，仅部分有用户取消操作的api才会用到。
     */
    cancel?: (res: any) => void;
  }

  interface MenuCallbacks extends Callbacks {
    /**
     * 监听Menu中的按钮点击时触发的方法，该方法仅支持Menu中的相关接口。
     */
    trigger?: (res: any) => void;
  }

  interface CheckJsApiOptions extends Callbacks {
    /**
     * 需要检测的JS接口列表
     */
    jsApiList: JSApis[];
  }

  interface OnMenuShareTimelineOptions extends MenuCallbacks {
    /**
     * 分享标题
     */
    title: string;

    /**
     * 分享链接，该链接域名或路径必须与当前页面对应的公众号JS安全域名一致
     */
    link: string;

    /**
     * 分享图标
     */
    imgUrl?: string;
  }

  interface OnMenuShareAppMessageOptions extends OnMenuShareTimelineOptions {
    /**
     * 分享描述
     */
    desc?: string;

    /**
     * 分享类型,music、video或link，不填默认为link
     */
    type?: 'music' | 'video' | 'link';

    /**
     * 如果type是music或video，则要提供数据链接，默认为空
     */
    dataUrl?: string;
  }

  interface OnMenuShareQQ extends OnMenuShareTimelineOptions {
    /**
     * 分享描述
     */
    desc?: string;
  }

  interface OnMenuShareWeibo extends OnMenuShareTimelineOptions {
    /**
     * 分享描述
     */
    desc?: string;
  }

  interface OnMenuShareQZone extends OnMenuShareTimelineOptions {
    /**
     * 分享描述
     */
    desc?: string;
  }

  interface ChooseImageOptions extends Callbacks {
    /**
     * 一次可选择图片数量
     *
     * 默认为 9
     */
    count?: number;

    /**
     * 可以指定是原图还是压缩图，默认二者都有
     */
    sizeType?: ImageSize[];

    /**
     * 可以指定来源是相册还是相机，默认二者都有
     */
    sourceType?: SourceType[];
  }

  interface PreviewImageOptions extends Callbacks {
    /**
     * 当前显示图片的 http 链接
     */
    current: string;

    /**
     * 需要预览的图片 http 链接列表
     */
    urls: string[];
  }

  interface UploadImageOptions extends Callbacks {
    /**
     * 需要上传的图片的本地ID，由chooseImage接口获得
     */
    localId: string;

    /**
     * 显示进度提示
     */
    isShowProgressTips?: boolean;
  }

  interface DownloadImageOptions extends Callbacks {
    /**
     * 需要下载的图片的服务器端ID，由uploadImage接口获得
     */
    serverId: string;

    /**
     * 显示进度提示
     */
    isShowProgressTips?: boolean;
  }

  interface GetLocalImgDataOptions extends Callbacks {
    /**
     * 图片的localID
     */
    localId: string;
  }
// ...
  interface ScanQRCodeOptions extends Callbacks {
    /**
     * 默认为0，扫描结果由微信处理，1则直接返回扫描结果
     */
    needResult?: 0 | 1;

    /**
     * 可以指定扫二维码还是一维码，默认二者都有
     */
    scanType?: scanTypes[];
  }
// ...
  interface ChooseWXPayOptions extends Callbacks {
    /**
     * 支付签名时间戳，注意微信jssdk中的所有使用timestamp字段均为小写。
     * 但最新版的支付后台生成签名使用的timeStamp字段名需大写其中的S字符
     */
    timestamp: string;

    /**
     * 支付签名随机串，不长于 32 位
     */
    nonceStr: string;

    /**
     * 统一支付接口返回的prepay_id参数值，提交格式如：prepay_id=***）
     */
    package: string;

    /**
     * 签名方式，默认为'SHA1'，使用新版支付需传入'MD5'
     */
    signType: 'SHA1' | 'MD5';

    /**
     * 支付签名
     */
    paySign: string;
  }
// ...
}